voxplot/requirements.txt
# VoxPlot Dependencies
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and data handling
PyYAML>=6.0
pathlib2>=2.3.0; python_version < "3.4"

# Machine Learning and AI Framework Dependencies
# Core ML libraries (required for ML analysis)
scikit-learn>=1.0.0      # Clustering, dimensionality reduction, regression
scikit-image>=0.19.0     # Image processing for spatial analysis

# Advanced ML libraries (optional but recommended)
# Uncomment based on your analysis needs
# umap-learn>=0.5.3      # UMAP dimensionality reduction
# xgboost>=1.6.0         # Gradient boosting for regression analysis
# lightgbm>=3.3.0        # Alternative gradient boosting

# GPU Acceleration Support (optional - for CUDA-enabled systems)
# IMPORTANT: cuML and CuPy must be installed via conda, not pip
# 
# Installation commands:
# conda install -c conda-forge -c rapidsai cuml cupy
# 
# For specific CUDA versions:
# conda install -c conda-forge -c rapidsai cuml cupy=cuda11x  # CUDA 11.x
# conda install -c conda-forge -c rapidsai cuml cupy=cuda12x  # CUDA 12.x
#
# DO NOT use pip install for these packages - they will fail to build

# Statistical analysis
statsmodels>=0.13.0      # Advanced statistical modeling

# Data export and I/O
openpyxl>=3.0.9          # Excel file support
xlsxwriter>=3.0.0        # Enhanced Excel writing capabilities

# Optional: Enhanced performance and additional features
# Uncomment if needed for large datasets
# numba>=0.56.0          # JIT compilation for performance
# dask>=2022.1.0         # Parallel computing for large datasets
# h5py>=3.6.0            # HDF5 support for large files
# netCDF4>=1.5.8         # NetCDF support
# xarray>=0.20.0         # Multi-dimensional labeled arrays

# Optional: Advanced visualization
# plotly>=5.0.0          # Interactive 3D visualizations
# bokeh>=2.4.0           # Interactive web-based visualizations

# Development and testing (optional)
# pytest>=6.2.0
# pytest-cov>=3.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.931